requirements.txt
groq
google-generativeai
openai
requests
python-dotenv
SpeechRecognition
pyaudio
elevenlabs
selenium
webdriver-manager
pyautogui
Pillow
rich
